[obey-campaign:8deed8b4] test(dungeon): split oversized dungeon integration test file (PR #368 review)
Addresses the review comment on tests/integration/dungeon_test.go (over
1,000 lines). Relocate all dungeon-move integration tests into
dungeon_move_test.go (with the move-only helpers), and the new dry-run
and batch tests into dungeon_move_batch_test.go. dungeon_test.go keeps
list/commands/intent/manifest coverage plus the shared setup helper.
Pure test relocation, no behavior change.
